requirements-doc.txt
# === Backend com FastAPI ===
fastapi==0.115.12
uvicorn==0.34.3

# === Requisições HTTP internas ===
requests==2.32.3

# === Variáveis de ambiente ===
python-dotenv==1.1.0

# === Leitura e extração de texto de arquivos PDF ===
PyMuPDF==1.26.0

# === Geração de PDF simples ===
fpdf==1.7.2

# === Integração com Gmail API (email_reader.py) ===
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-python-client==2.171.0

# === Gemini / Google Generative AI ===
google-generativeai==0.5.4

# === Testes ===
pytest==8.4.0
pytest-asyncio==1.0.0
